+++ Models/TransferRequest.cs
@@ -5,6 +5,7 @@ public sealed class TransferRequest {
 	public required string SourceBotName { get; init; }
 	public required string TargetBotName { get; init; }
 	public required List<string> Modes { get; init; }
+	public required bool Force { get; init; }
 	public required int WhitelistedUniqueItemCount { get; init; }
 	public required List<TransferBatch> Batches { get; init; }